+++ lib/src/platform_io.dart
@@ -1,0 +1,1 @@
+export 'platform_io_stub.dart' if (dart.library.io) 'platform_io_vm.dart';